Restructured QuizPage, QuizResultsPage, QuizCard, and QuizResultCard. Added /quiz to Navigation bar training tab route
