+++ react-app/src/components/NavigationBar/NavigationBar.tsx
@@ -64,7 +64,7 @@ function NavigationBar() {
           <NavLink
             to="/trainings"
             className={({ isActive }) =>
-              isActive || location === "/trainingpage"
+              isActive || location === "/trainingpage" || location === "/quiz"
                 ? `${styles.tab} ${styles.tabActive}`
                 : `${styles.tab} ${styles.tabInActive}`
             }
