--- react-app/src/pages/QuizPage/QuizComponent/QuizComponent.module.css
@@ -1,6 +1,0 @@
-.container {
-  background-color: var(--blue-purple);
-  padding: 2%;
-  border-radius: 2%;
-  margin-top: 3%;
-}
