+++ react-app/src/pages/QuizResultPage/QuizResultCard/QuizResultCard.module.css
