+++ react-app/src/pages/QuizResultPage/QuizResultCard/QuizResultCard.tsx
@@ -8,7 +8,7 @@ import {
   grayRadioButton,
   greenRadioButton,
   redRadioButton,
-} from "../../../../muiTheme";
+} from "../../../muiTheme";
 import styles from "./QuizResultCard.module.css";
 import { IoCloseOutline, IoCheckmark } from "react-icons/io5";
